requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
spacy>=3.2.0
gensim>=4.1.2
transformers>=4.15.0
nltk>=3.6.5
requests>=2.26.0
beautifulsoup4>=4.10.0
fastapi>=0.70.0
uvicorn>=0.15.0
pytest>=6.2.5
python-dotenv>=0.19.2 